Building and running, but display currently does not work.

+++ SpriteBuilder/CCButton/CCBPButton.m
@@ -238,52 +238,52 @@
 
 - (void) setLabelColor:(CCColor*)color forState:(CCBPControlState)state
 {
-    [_labelColors setObject:color forKey:[NSNumber numberWithInt:state]];
+    [_labelColors setObject:color forKey:[NSNumber numberWithInt:(int)state]];
     [self stateChanged];
 }
 
 - (CCColor*) labelColorForState:(CCBPControlState)state
 {
-    CCColor* color = [_labelColors objectForKey:[NSNumber numberWithInt:state]];
+    CCColor* color = [_labelColors objectForKey:[NSNumber numberWithInt:(int)state]];
     if (!color) color = [CCColor whiteColor];
     return color;
 }
 
 - (void) setLabelOpacity:(CGFloat)opacity forState:(CCBPControlState)state
 {
-    [_labelOpacities setObject:[NSNumber numberWithFloat:opacity] forKey:[NSNumber numberWithInt:state]];
+    [_labelOpacities setObject:[NSNumber numberWithFloat:opacity] forKey:[NSNumber numberWithInt:(int)state]];
     [self stateChanged];
 }
 
 - (CGFloat) labelOpacityForState:(CCBPControlState)state
 {
-    NSNumber* val = [_labelOpacities objectForKey:[NSNumber numberWithInt:state]];
+    NSNumber* val = [_labelOpacities objectForKey:[NSNumber numberWithInt:(int)state]];
     if (!val) return 1;
     return [val floatValue];
 }
 
 - (void) setBackgroundColor:(CCColor*)color forState:(CCBPControlState)state
 {
-    [_backgroundColors setObject:color forKey:[NSNumber numberWithInt:state]];
+    [_backgroundColors setObject:color forKey:[NSNumber numberWithInt:(int)state]];
     [self stateChanged];
 }
 
 - (CCColor*) backgroundColorForState:(CCBPControlState)state
 {
-    CCColor* color = [_backgroundColors objectForKey:[NSNumber numberWithInt:state]];
+    CCColor* color = [_backgroundColors objectForKey:[NSNumber numberWithInt:(int)state]];
     if (!color) color = [CCColor whiteColor];
     return color;
 }
 
 - (void) setBackgroundOpacity:(CGFloat)opacity forState:(CCBPControlState)state
 {
-    [_backgroundOpacities setObject:[NSNumber numberWithFloat:opacity] forKey:[NSNumber numberWithInt:state]];
+    [_backgroundOpacities setObject:[NSNumber numberWithFloat:opacity] forKey:[NSNumber numberWithInt:(int)state]];
     [self stateChanged];
 }
 
 - (CGFloat) backgroundOpacityForState:(CCBPControlState)state
 {
-    NSNumber* val = [_backgroundOpacities objectForKey:[NSNumber numberWithInt:state]];
+    NSNumber* val = [_backgroundOpacities objectForKey:[NSNumber numberWithInt:(int)state]];
     if (!val) return 1;
     return [val floatValue];
 }
@@ -292,18 +292,18 @@
 {
     if (spriteFrame)
     {
-        [_backgroundSpriteFrames setObject:spriteFrame forKey:[NSNumber numberWithInt:state]];
+        [_backgroundSpriteFrames setObject:spriteFrame forKey:[NSNumber numberWithInt:(int)state]];
     }
     else
     {
-        [_backgroundSpriteFrames removeObjectForKey:[NSNumber numberWithInt:state]];
+        [_backgroundSpriteFrames removeObjectForKey:[NSNumber numberWithInt:(int)state]];
     }
     [self stateChanged];
 }
 
 - (CCSpriteFrame*) backgroundSpriteFrameForState:(CCBPControlState)state
 {
-    return [_backgroundSpriteFrames objectForKey:[NSNumber numberWithInt:state]];
+    return [_backgroundSpriteFrames objectForKey:[NSNumber numberWithInt:(int)state]];
 }
 
 - (void) setTitle:(NSString *)title
